Project is now a pure Perl implementation with leftover C code
This is a design decision I hesitate to make. I want to implement most
of this in C in order to grasp a deeper understanding of the C language,
but Perl (And Python, as an honorable mention) makes it so easy to bind
to C functions and libraries through its already-available modules.

In the future, I will have to decide what to do with this multi-language
implementation. Should I scrap C? Should I ditch Perl? Should I use a
mix of both? Or maybe I should split them into their own projects.

One thing is certain: Time is not a problem. This decision can be made
at any point in the future.

Signed-off-by: zzApotheosis <zzApotheosis@gmail.com>