Minor update
I added a small explanation for how the functionality of mutex locking
could have been private to the implementation. I also hid the
implementation details of the counter_t struct in the counter
implementation because the details of its internal data structure does
not need to be in its header.