Updated Emacs config
As much as I love Vim bindings, I'm slowly converting to the holy church of Emacs.

I'll keep Evil mode installed whenever I want to use it (the more the merrier) but I'll
default to having it disabled. It will be available any time I wish to enable it manually.