Updated C code for Rust C FFI example
I learned how to further abstract C source implementation
from my time at GaTech. It turns out, the header can simply
declare a struct type and typedef it and the source file
can define what's actually in the struct so long as the
function signatures in the header use ONLY a pointer to the
struct declaration.